+++ src/main/java/stirling/software/SPDF/model/ApplicationProperties.java
@@ -13,6 +13,10 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.PropertySource;
 
 import stirling.software.SPDF.config.YamlPropertySourceFactory;
+import stirling.software.SPDF.model.provider.GithubProvider;
+import stirling.software.SPDF.model.provider.GoogleProvider;
+import stirling.software.SPDF.model.provider.KeycloakProvider;
+import stirling.software.SPDF.model.provider.UnsupportedProviderException;
 
 @Configuration
 @ConfigurationProperties(prefix = "")
@@ -128,6 +132,15 @@ public class ApplicationProperties {
         private OAUTH2 oauth2;
         private int loginAttemptCount;
         private long loginResetTimeMinutes;
+        private String loginMethod = "all";
+
+        public String getLoginMethod() {
+            return loginMethod;
+        }
+
+        public void setLoginMethod(String loginMethod) {
+            this.loginMethod = loginMethod;
+        }
 
         public int getLoginAttemptCount() {
             return loginAttemptCount;
@@ -187,6 +200,8 @@ public class ApplicationProperties {
                     + initialLogin
                     + ", csrfDisabled="
                     + csrfDisabled
+                    + ", loginMethod="
+                    + loginMethod
                     + "]";
         }
 
@@ -355,7 +370,7 @@ public class ApplicationProperties {
                 private GithubProvider github = new GithubProvider();
                 private KeycloakProvider keycloak = new KeycloakProvider();
 
-                public Provider get(String registrationId) throws Exception {
+                public Provider get(String registrationId) throws UnsupportedProviderException {
                     switch (registrationId.toLowerCase()) {
                         case "google":
                             return getGoogle();
@@ -366,7 +381,8 @@ public class ApplicationProperties {
                         default:
                             break;
                     }
-                    throw new Exception("Provider not supported, use custom setting.");
+                    throw new UnsupportedProviderException(
+                            "Logout from the provider is not supported? Report it at https://github.com/Stirling-Tools/Stirling-PDF/issues");
                 }
 
                 public GoogleProvider getGoogle() {
@@ -404,311 +420,6 @@ public class ApplicationProperties {
                             + "]";
                 }
             }
-        }
-    }
-
-    public static class GoogleProvider extends Provider {
-
-        private static final String authorizationUri =
-                "https://accounts.google.com/o/oauth2/v2/auth";
-        private static final String tokenUri = "https://www.googleapis.com/oauth2/v4/token";
-        private static final String userInfoUri =
-                "https://www.googleapis.com/oauth2/v3/userinfo?alt=json";
-
-        public String getAuthorizationuri() {
-            return authorizationUri;
-        }
-
-        public String getTokenuri() {
-            return tokenUri;
-        }
-
-        public String getUserinfouri() {
-            return userInfoUri;
-        }
-
-        private String clientId;
-        private String clientSecret;
-        private Collection<String> scopes = new ArrayList<>();
-        private String useAsUsername = "email";
-
-        @Override
-        public String getClientId() {
-            return this.clientId;
-        }
-
-        @Override
-        public void setClientId(String clientId) {
-            this.clientId = clientId;
-        }
-
-        @Override
-        public String getClientSecret() {
-            return this.clientSecret;
-        }
-
-        @Override
-        public void setClientSecret(String clientSecret) {
-            this.clientSecret = clientSecret;
-        }
-
-        @Override
-        public Collection<String> getScopes() {
-            if (scopes == null || scopes.isEmpty()) {
-                scopes = new ArrayList<>();
-                scopes.add("https://www.googleapis.com/auth/userinfo.email");
-                scopes.add("https://www.googleapis.com/auth/userinfo.profile");
-            }
-            return scopes;
-        }
-
-        @Override
-        public void setScopes(String scopes) {
-            this.scopes =
-                    Arrays.stream(scopes.split(",")).map(String::trim).collect(Collectors.toList());
-        }
-
-        @Override
-        public String getUseAsUsername() {
-            return this.useAsUsername;
-        }
-
-        @Override
-        public void setUseAsUsername(String useAsUsername) {
-            this.useAsUsername = useAsUsername;
-        }
-
-        @Override
-        public String toString() {
-            return "Google [clientId="
-                    + clientId
-                    + ", clientSecret="
-                    + (clientSecret != null && !clientSecret.isEmpty() ? "MASKED" : "NULL")
-                    + ", scopes="
-                    + scopes
-                    + ", useAsUsername="
-                    + useAsUsername
-                    + "]";
-        }
-
-        @Override
-        public String getName() {
-            return "google";
-        }
-
-        @Override
-        public String getClientName() {
-            return "Google";
-        }
-
-        public boolean isSettingsValid() {
-            return super.isValid(this.getClientId(), "clientId")
-                    && super.isValid(this.getClientSecret(), "clientSecret")
-                    && super.isValid(this.getScopes(), "scopes")
-                    && isValid(this.getUseAsUsername(), "useAsUsername");
-        }
-    }
-
-    public static class GithubProvider extends Provider {
-        private static final String authorizationUri = "https://github.com/login/oauth/authorize";
-        private static final String tokenUri = "https://github.com/login/oauth/access_token";
-        private static final String userInfoUri = "https://api.github.com/user";
-
-        public String getAuthorizationuri() {
-            return authorizationUri;
-        }
-
-        public String getTokenuri() {
-            return tokenUri;
-        }
-
-        public String getUserinfouri() {
-            return userInfoUri;
-        }
-
-        private String clientId;
-        private String clientSecret;
-        private Collection<String> scopes = new ArrayList<>();
-        private String useAsUsername = "login";
-
-        @Override
-        public String getIssuer() {
-            return new String();
-        }
-
-        @Override
-        public void setIssuer(String issuer) {}
-
-        @Override
-        public String getClientId() {
-            return this.clientId;
-        }
-
-        @Override
-        public void setClientId(String clientId) {
-            this.clientId = clientId;
-        }
-
-        @Override
-        public String getClientSecret() {
-            return this.clientSecret;
-        }
-
-        @Override
-        public void setClientSecret(String clientSecret) {
-            this.clientSecret = clientSecret;
-        }
-
-        @Override
-        public Collection<String> getScopes() {
-            if (scopes == null || scopes.isEmpty()) {
-                scopes = new ArrayList<>();
-                scopes.add("read:user");
-            }
-            return scopes;
-        }
-
-        @Override
-        public void setScopes(String scopes) {
-            this.scopes =
-                    Arrays.stream(scopes.split(",")).map(String::trim).collect(Collectors.toList());
-        }
-
-        @Override
-        public String getUseAsUsername() {
-            return this.useAsUsername;
-        }
-
-        @Override
-        public void setUseAsUsername(String useAsUsername) {
-            this.useAsUsername = useAsUsername;
-        }
-
-        @Override
-        public String toString() {
-            return "GitHub [clientId="
-                    + clientId
-                    + ", clientSecret="
-                    + (clientSecret != null && !clientSecret.isEmpty() ? "MASKED" : "NULL")
-                    + ", scopes="
-                    + scopes
-                    + ", useAsUsername="
-                    + useAsUsername
-                    + "]";
-        }
-
-        @Override
-        public String getName() {
-            return "github";
-        }
-
-        @Override
-        public String getClientName() {
-            return "GitHub";
-        }
-
-        public boolean isSettingsValid() {
-            return super.isValid(this.getClientId(), "clientId")
-                    && super.isValid(this.getClientSecret(), "clientSecret")
-                    && super.isValid(this.getScopes(), "scopes")
-                    && isValid(this.getUseAsUsername(), "useAsUsername");
-        }
-    }
-
-    public static class KeycloakProvider extends Provider {
-        private String issuer;
-        private String clientId;
-        private String clientSecret;
-        private Collection<String> scopes = new ArrayList<>();
-        private String useAsUsername = "email";
-
-        @Override
-        public String getIssuer() {
-            return this.issuer;
-        }
-
-        @Override
-        public void setIssuer(String issuer) {
-            this.issuer = issuer;
-        }
-
-        @Override
-        public String getClientId() {
-            return this.clientId;
-        }
-
-        @Override
-        public void setClientId(String clientId) {
-            this.clientId = clientId;
-        }
-
-        @Override
-        public String getClientSecret() {
-            return this.clientSecret;
-        }
-
-        @Override
-        public void setClientSecret(String clientSecret) {
-            this.clientSecret = clientSecret;
-        }
-
-        @Override
-        public Collection<String> getScopes() {
-            if (scopes == null || scopes.isEmpty()) {
-                scopes = new ArrayList<>();
-                scopes.add("profile");
-                scopes.add("email");
-            }
-            return scopes;
-        }
-
-        @Override
-        public void setScopes(String scopes) {
-            this.scopes =
-                    Arrays.stream(scopes.split(",")).map(String::trim).collect(Collectors.toList());
-        }
-
-        @Override
-        public String getUseAsUsername() {
-            return this.useAsUsername;
-        }
-
-        @Override
-        public void setUseAsUsername(String useAsUsername) {
-            this.useAsUsername = useAsUsername;
-        }
-
-        @Override
-        public String toString() {
-            return "Keycloak [issuer="
-                    + issuer
-                    + ", clientId="
-                    + clientId
-                    + ", clientSecret="
-                    + (clientSecret != null && !clientSecret.isEmpty() ? "MASKED" : "NULL")
-                    + ", scopes="
-                    + scopes
-                    + ", useAsUsername="
-                    + useAsUsername
-                    + "]";
-        }
-
-        @Override
-        public String getName() {
-            return "keycloak";
-        }
-
-        @Override
-        public String getClientName() {
-            return "Keycloak";
-        }
-
-        public boolean isSettingsValid() {
-            return isValid(this.getIssuer(), "issuer")
-                    && isValid(this.getClientId(), "clientId")
-                    && isValid(this.getClientSecret(), "clientSecret")
-                    && isValid(this.getScopes(), "scopes")
-                    && isValid(this.getUseAsUsername(), "useAsUsername");
         }
     }
