+++ src/main/java/stirling/software/SPDF/model/provider/UnsupportedProviderException.java
@@ -1,0 +1,7 @@
+package stirling.software.SPDF.model.provider;
+
+public class UnsupportedProviderException extends Exception {
+    public UnsupportedProviderException(String message) {
+        super(message);
+    }
+}
