+++ src/main/resources/templates/login.html
@@ -114,7 +114,7 @@
           <img class="mb-4" src="favicon.svg" alt="favicon" width="144" height="144">
 
           <h1 class="h1 mb-3 fw-normal" th:text="${@appName}">Stirling-PDF</h1>
-          <div th:if="${oAuth2Enabled}">
+          <div th:if="${oAuth2Enabled} and (${loginMethod} == 'all' or ${loginMethod} == 'oauth2')">
             <a href="#" class="w-100 btn btn-lg btn-primary" data-bs-toggle="modal" data-bs-target="#loginsModal" th:text="#{login.ssoSignIn}">Login Via SSO</a>
             <br>
             <br>
@@ -134,7 +134,7 @@
             </div>
           </div>
         </div>
-        <form th:action="@{login}" method="post">
+        <form th:if="${loginMethod} == 'all' or ${loginMethod} == 'normal'" th:action="@{login}" method="post">
           <h2 class="h5 mb-3 fw-normal" th:text="#{login.signinTitle}">Please sign in</h2>
           <div class="form-floating">
             <input type="text" class="form-control" id="username" name="username" placeholder="admin">
